Implement an exercise varnishtest can test
Moves exercises into exercises/*.test, which is mostly RST-like syntax. We
then extract different parts of it to build the complete exercise. This
gives us flexibility (e.g: We can stick the solution at the end of the PDF
but in-line for slides) and the ability to verify that our exercises
actually work as expected.

Work in progress.