Use the pre-installed CGI script
What we want to test here is Varnish, so instead of suggesting CGI as a
solution it should be part of the material. In our training VMs we
currently ship a `sleep` script that does what we want on the backend
side.

This is necessary especially since this is not a course on CGI, and the
first explanation on how to set it up happens five chapters after [1]
this exercise.

My suggestion is to include the scripts to the training material,
instruct trainees to copy it to the proper location, and show its
contents [1] on the exercice slide. That would be for every single CGI
script, we always keep them short.

[1] currently: 10.3.2 Exercise: Grace
[2] using the include directive to stay in sync